+++ ChangeLog.md
@@ -1,6 +1,7 @@
 ## 0.6.1
 
 * Switch to V2 upload by default
+* --skip-hoogle option
 
 ## 0.6.0